Missing dashes before dapr-http-port
The command for running the application using the Dapr CLI is not formatted correctly. The argument dapr-http-port is missing two dashes. Without these dashes the dapr cli try to run an executable file called dapr-http-port which is not the intention of this command. Adding two dashes before the argument resolve the issue.